pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uniscale"
version = "0.1.0"
description = "Research project for scale-invariant tokenizer selection using small language models"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Saibo", email = "saibo@example.com"}
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "datasets>=2.14.0",
    "accelerate>=0.24.0",
    "tokenizers>=0.15.0",
    "sentencepiece>=0.1.99",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
    "pyyaml>=6.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["uniscale*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100
